requirements.txt
flask
python-box
pyyaml
cherrypy
Flask-Session
